backend/requirements.txt
django
djangorestframework
channels
channels-redis
playwright
pydantic
dramatiq
redis
google-generativeai